% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomes.R
\docType{data}
\name{.ASSEMBLY_REPORT_COLS}
\alias{.ASSEMBLY_REPORT_COLS}
\title{Column names of an NCBI assembly report}
\format{
An object of class \code{character} of length 10.
}
\usage{
.ASSEMBLY_REPORT_COLS
}
\description{
Column names of an NCBI assembly report
}
\keyword{internal}
