% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotation.R
\name{.decompress_gz}
\alias{.decompress_gz}
\title{Streaming Gzip Decompressor}
\usage{
.decompress_gz(gz_file, dest_file)
}
\arguments{
\item{gz_file}{Path to input .gz file.}

\item{dest_file}{Path to output decompressed file.}
}
\value{
The destination file path invisibly.
}
\description{
Streaming Gzip Decompressor
}
\keyword{internal}
