% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotation.R
\name{.getAnnotationNCBICoordsGFF}
\alias{.getAnnotationNCBICoordsGFF}
\title{Get NCBI Genomic Coordinates from GFF3}
\usage{
.getAnnotationNCBICoordsGFF(
  species,
  annotation_dir,
  force_update = FALSE,
  offline = getOption("geneRosetta.offline", FALSE)
)
}
\arguments{
\item{species}{Species name.}

\item{annotation_dir}{Directory path.}

\item{force_update}{Logical.}

\item{offline}{Logical.}
}
\description{
Get NCBI Genomic Coordinates from GFF3
}
\keyword{internal}
