% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomes.R
\docType{data}
\name{.NCBI_GENOMES_URL}
\alias{.NCBI_GENOMES_URL}
\title{Base URL of the NCBI genomes FTP tree}
\format{
An object of class \code{character} of length 1.
}
\usage{
.NCBI_GENOMES_URL
}
\description{
Base URL of the NCBI genomes FTP tree
}
\keyword{internal}
