% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneRosetta-package.R
\docType{package}
\name{geneRosetta-package}
\alias{geneRosetta}
\alias{geneRosetta-package}
\title{geneRosetta: Gene Annotation Retrieval and Mapping Utilities}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Reliable Gene Annotation Mapping. Retrieve, cache, and merge human, mouse, and rat gene annotations from Ensembl BioMart, NCBI, and organism-specific authorities (HGNC, MGI, RGD); map between gene symbols and stable identifiers with synonym-aware matching and configurable multi-mapping resolution.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexankc/geneRosetta}
  \item Report bugs at \url{https://github.com/alexankc/geneRosetta/issues}
}

}
\author{
\strong{Maintainer}: Alexandros Karagiannopoulos \email{alexandros.karagiannopoulos@med.lu.se}

}
\keyword{internal}
